+++ components/MotionSorter.php
@@ -188,6 +188,32 @@ class MotionSorter
         return $motionsSorted;
     }
 
+    private static function getSortId(IMotion $imotion): string
+    {
+        if (is_a($imotion, Motion::class)) {
+            return 'motion-' . $imotion->id;
+        } else {
+            return 'amendment-' . $imotion->id;
+        }
+    }
+
+    /**
+     * @param string[] $usedKeys
+     */
+    private static function getUnusedSortKey(IMotion $imotion, array $usedKeys): string
+    {
+        if (!in_array($imotion->titlePrefix, $usedKeys)) {
+            return (string)$imotion->titlePrefix; // Hint: for amendments, can still be null at the moment
+        }
+
+        $i = 1;
+        while (in_array($imotion->titlePrefix . ' (' . $i . ')', $usedKeys)) {
+            $i++;
+        }
+
+        return $imotion->titlePrefix . ' (' . $i . ')';
+    }
+
     /**
      * @param IMotion[] $imotions
      * @return array<string, array<string, IMotion>>
@@ -199,8 +225,7 @@ class MotionSorter
         $motionIdsToBeSorted = [];
 
         foreach ($imotions as $imotion) {
-            $motionIdsToBeSorted[] = $imotion->id;
-            // @TODO A differenciation between motions and amendments will be necessary
+            $motionIdsToBeSorted[] = self::getSortId($imotion);
         }
 
         $statuses = $consultation->getStatuses()->getInvisibleMotionStatuses();
@@ -208,7 +233,7 @@ class MotionSorter
         foreach ($items as $agendaItem) {
             $agendaMotions = $agendaItem->getMyIMotions(IMotionStatusFilter::onlyUserVisible($consultation, true));
             foreach ($agendaMotions as $agendaMotion) {
-                if (!in_array($agendaMotion->id, $motionIdsToBeSorted)) {
+                if (!in_array(self::getSortId($agendaMotion), $motionIdsToBeSorted)) {
                     continue;
                 }
                 if (!self::imotionIsVisibleOnHomePage($agendaMotion, $statuses)) {
@@ -217,17 +242,18 @@ class MotionSorter
                 if (!isset($motionsSorted['agenda' . $agendaItem->id])) {
                     $motionsSorted['agenda' . $agendaItem->id] = [];
                 }
-                $key                                             = $agendaMotion->titlePrefix;
+                $key = self::getUnusedSortKey($agendaMotion, array_keys($motionsSorted['agenda' . $agendaItem->id]));
                 $motionsSorted['agenda' . $agendaItem->id][$key] = $agendaMotion;
-                $foundMotionIds[]                                = $agendaMotion->id;
+                $foundMotionIds[] = self::getSortId($agendaMotion);
             }
         }
         foreach ($imotions as $motion) {
-            if (!in_array($motion->id, $foundMotionIds)) {
+            if (!in_array(self::getSortId($motion), $foundMotionIds)) {
                 if (!isset($motionsSorted['noAgenda'])) {
                     $motionsSorted['noAgenda'] = [];
                 }
-                $motionsSorted['noAgenda'][$motion->titlePrefix] = $motion;
+                $key = self::getUnusedSortKey($motion, array_keys($motionsSorted['noAgenda']));
+                $motionsSorted['noAgenda'][$key] = $motion;
             }
         }