+++ config/defines.php
@@ -1,8 +1,8 @@
 <?php
 
-define('ANTRAGSGRUEN_VERSION', '4.15.3');
+define('ANTRAGSGRUEN_VERSION', '4.16.1');
 define('ANTRAGSGRUEN_HISTORY_URL', 'https://github.com/CatoTH/antragsgruen/blob/main/History.md');
 define('ANTRAGSGRUEN_UPDATE_BASE', 'https://antragsgruen.de/updates/');
 define('ANTRAGSGRUEN_MIN_PHP_VERSION', '8.1.0');
-define('ANTRAGSGRUEN_NEXT_PHP_VERSION', '8.1.0');
+define('ANTRAGSGRUEN_NEXT_PHP_VERSION', '8.2.0');
 define('ANTRAGSGRUEN_INSTALLATION_SOURCE', 'repository-source');