+++ docs/create-dist.sh
@@ -89,6 +89,7 @@ rm migrations/m1*
 rm migrations/m20*
 rm migrations/m21*
 rm migrations/m22*
+rm migrations/m23*
 find . -name ".DS_Store" -exec rm {} \;
 find . -name ".git" -exec rm -rf {} \;
 find . -name ".github" -exec rm -rf {} \;
@@ -132,6 +133,7 @@ rm phpstan.use-baseline.neon
 rm eslint.config.mjs
 rm .eslintrc.cjs
 rm .eslintrc.js
+rm docker-compose.development.yml
 
 cd ..
 tar cfj ../dist/antragsgruen-$ANTRAGSGRUEN_VERSION.tar.bz2 antragsgruen-$ANTRAGSGRUEN_VERSION