+++ models/db/AmendmentProposal.php
@@ -5,10 +5,12 @@ declare(strict_types=1);
 namespace app\models\db;
 
 use app\components\UrlHelper;
+use app\models\proposedProcedure\Agenda;
 use app\models\settings\AntragsgruenApp;
 use app\components\diff\AmendmentCollissionDetector;
 use app\models\exceptions\Internal;
 use app\models\sectionTypes\ISectionType;
+use yii\behaviors\AttributeTypecastBehavior;
 
 /**
  * @property int $amendmentId
@@ -53,7 +55,7 @@ class AmendmentProposal extends IProposal
         if ($includeOtherAmendments && $this->proposalStatus === Amendment::STATUS_OBSOLETED_BY_AMENDMENT) {
             $obsoletedBy = $consultation->getAmendment(intval($this->comment));
             if ($obsoletedBy && $internalNestingLevel < 10) {
-                return $obsoletedBy->getLatestProposal()->hasAlternativeProposaltext($includeOtherAmendments, $internalNestingLevel + 1);
+                return $obsoletedBy->getLatestProposal(true)->hasAlternativeProposaltext($includeOtherAmendments, $internalNestingLevel + 1);
             }
         }
 
@@ -74,6 +76,43 @@ class AmendmentProposal extends IProposal
             [['amendmentId'], 'required'],
             [['amendmentId'], 'number'],
         ];
+    }
+
+    public function behaviors(): array
+    {
+        return [
+            'typecast' => [
+                'class' => AttributeTypecastBehavior::class,
+                'attributeTypes' => [
+                    'id' => AttributeTypecastBehavior::TYPE_INTEGER,
+                ],
+                'typecastAfterValidate' => true,
+                'typecastBeforeSave' => true,
+                'typecastAfterFind' => true,
+            ],
+        ];
+    }
+
+    public function flushViewCaches(): void
+    {
+        $amendment = $this->getAmendment();
+        if ($amendment?->getMyMotion()) {
+            $amendment->getMyMotion()->flushViewCache();
+            Agenda::getProposedAmendmentProcedureCache($amendment, $this)->flushCache();
+        }
+    }
+
+    /**
+     * @param bool $runValidation
+     * @param null $attributeNames
+     *
+     * @return bool
+     */
+    public function save($runValidation = true, $attributeNames = null)
+    {
+        $result = parent::save($runValidation, $attributeNames);
+        $this->flushViewCaches();
+        return $result;
     }
 
     public function getAmendment(): ?Amendment
@@ -139,7 +178,7 @@ class AmendmentProposal extends IProposal
         if ($this->proposalStatus === Amendment::STATUS_OBSOLETED_BY_AMENDMENT) {
             $obsoletedBy = $this->getMyConsultation()->getAmendment(intval($this->comment));
             if ($obsoletedBy && $internalNestingLevel < 10) {
-                return $obsoletedBy->getLatestProposal()->getAlternativeProposaltextReference($internalNestingLevel + 1);
+                return $obsoletedBy->getLatestProposal(true)->getAlternativeProposaltextReference($internalNestingLevel + 1);
             }
         }