+++ plugins/green_manager/views/manager/index.php
@@ -17,6 +17,30 @@ $controller  = $this->context;
 $controller->layoutParams->addCSS('css/manager.css');
 $controller->layoutParams->canonicalUrl      = 'https://discuss.green/';
 $controller->layoutParams->alternateLanuages = ['de' => 'https://antragsgruen.de/'];
+$controller->layoutParams->addInlineCss('
+    .homeFigure { text-align: center; }
+    .homeFigure figcaption { margin-top: -20px; margin-bottom: 20px; font-size: 0.8em; font-style: italic; }
+    .homeFigureAmendment img { max-width: 100%; }
+    @media (min-width: 800px) {
+        .homeFigureAmendment img { max-width: 600px; }
+    }
+    .homeFigurePrint { max-width: 230px; box-shadow: 0 0 7px rgba(0,0,0,.4); border-radius: 2px; overflow: hidden; }
+    @media (min-width: 800px) {
+        .homeFigurePrint { float: right; margin-left: 50px; }
+    }
+    @media (max-width: 799px) {
+        .homeFigurePrint { margin: 20px auto; }
+    }
+    .homeFigurePrint img { max-width: 100%; }
+    .homeFigurePrint figcaption { margin-bottom: 5px; }
+
+    @media (min-width: 800px) {
+        .homeFigureTwoHolder { display: flex; flex-direction: row; margin-bottom: 30px; margin-top: -25px; }
+        .homeFigureTwoHolder > * { flex-basis: 50%; }
+    }
+    .homeFigureTwoHolder img { max-width: 100%; }
+    .homeFigureSpeech figcaption { margin-top: -10px; }
+');
 
 $layout                   = $controller->layoutParams;
 $layout->bodyCssClasses[] = 'manager_index';
@@ -31,98 +55,183 @@ $params = \app\models\settings\AntragsgruenApp::getInstance();
 
 <div class="content infoSite">
     <p>
-        Discuss.green offers a clear and efficient tool for the administration of motions, resolutions, amendments and candidacies,
-        specifically designed for conventions and congresses of parties, social initiatives and organizations.
+        Antragsgrün is an easy-to-use online tool for NGOs, political parties, and social initiatives to collaboratively discuss resolutions, party platforms, and amendments.<br>
+        It helps to manage candidacies and supports meetings by providing online votings, speaking lists, and many more features.
     </p>
 
     <p>
-        It has been created for managing both national and regional congresses of the German Greens
-        and is by now used by a variety of organizations within and beyond the green party.
+        It was originally created for managing regional congresses of the German Greens, but was quickly adopted
+        on the Federal and European level, as well as by organizations beyond the family of Green parties.
     </p>
 
     <p>
-        It's available as open source and we provide free hosting for all green parties.
+        It's available as open source and we provide free hosting for Green parties.
     </p>
 </div>
 
-
-<h2 id="who_uses_it" class="green">Who is already using it?</h2>
-<div class="content contentWhoUsesIt">
-    <div class="list">
-        <a href="https://motiontool.europeangreens.eu/" target="_blank">
-            <img src="<?= $assets->baseUrl ?>/logo-egp.svg" alt="European Greens">
-            <div class="name">European Greens</div>
-            <div class="hint">EGP Online Congress</div>
-        </a>
-        <a href="https://antraege.gruene.de/" target="_blank">
-            <img src="<?= $assets->baseUrl ?>/logo-b90.svg" alt="Bündnis 90 / Die GRÜNEN">
-            <div class="name">Bündnis 90 / Die GRÜNEN</div>
-            <div class="hint">Federal Convention</div>
-        </a>
+<section aria-labelledby="who_uses_it">
+    <h2 id="who_uses_it" class="green">Who is already using it?</h2>
+    <div class="content contentWhoUsesIt">
+        <div class="list">
+            <a href="https://motiontool.europeangreens.eu/" target="_blank">
+                <img src="<?= $assets->baseUrl ?>/logo-egp.svg" alt="European Greens">
+                <div class="name">European Greens</div>
+                <div class="hint">EGP Online Congress</div>
+            </a>
+            <a href="https://antraege.gruene.de/" target="_blank">
+                <img src="<?= $assets->baseUrl ?>/logo-b90.svg" alt="Bündnis 90 / Die GRÜNEN">
+                <div class="name">Bündnis 90 / Die GRÜNEN</div>
+                <div class="hint">Federal Convention</div>
+            </a>
+        </div>
+        <p>
+            ... as well as dozens of other green organizations (like the Greens in
+            <a href="https://gruene-wien.antragsgruen.de/" target="_blank">Vienna</a> and
+            <a href="https://edinburghgreens.discuss.green/" target="_blank">Edinburgh</a>,
+            non-green parties like the Austrian NEOS or the Dutch Volt,
+            and civil organizations like the <a href="https://yfj-votes.motion.tools/" target="_blank">European Youth Forum</a> or the
+            <a href="https://ga.cdnee.org" target="_blank">Cooperation &amp; Developement Network Eastern Europe</a>.
+        </p>
     </div>
-    <p>
-        ... as well as dozens of other green organizations (like the Greens in
-        <a href="https://gruene-wien.antragsgruen.de/" target="_blank">Vienna</a> and
-        <a href="https://edinburghgreens.discuss.green/" target="_blank">Edinburgh</a>,
-        non-green parties like the Austrian NEOS or the Dutch Volt,
-        and civil organizations like the <a href="https://yfj-votes.motion.tools/" target="_blank">European Youth Forum</a> or the
-        <a href="https://tooldoku.dbjr.de/category/antragsgruen/" target="_blank">German Federal Youth Council</a>.
-    </p>
-</div>
+</section>
 
-<h2 id="funktionen" class="green">Core functionality</h2>
-<div class="content infoSite">
-    <ul>
-        <li><strong>Submit motions, resolutions, discussion papers and applications online</strong><br>
-            Flexible and user-friendly submission of motions and applications<br>
-            Intuitive creation of amendments<br>
-            Creating final resolutions based on the motions and amendments to them.<br>
-            Managing the agenda of a congress
-        </li>
-        <li><strong>Discuss motions</strong><br>
-            Motions as well as amendments can be commented upon, either as a whole document or per paragraph. Depending
-            on preference, the comments function can be open to everybody or restricted to registered users.<br>
-            <small>(No comments required? The comments function can be easily deactivated.)</small>
-        </li>
-        <li><strong>Sophisticated administration tools</strong><br>
-            Filter and sorting options for all motions / amendments<br>
-            If required: assessment of each submitted motion / amendment with regards to permissibility by the programme
-            commission prior to publication.<br>
-            The page visibility can be specified: from “openly visible” to “only for invited members”.<br>
-            All user interface texts can be adapted.<br>
-            Flexible layout: the site provides different layout variants and the logos are exchangeable. If required, we
-            can adapt the design to your CI specifications.
-        </li>
-        <li><strong>Diverse export options and notifications</strong><br>
-            Automatically generated PDFs for all motions and amendments with different templates<br>
-            Export to different office formats (OpenDocument text, spreadsheet)<br>
-            E-Mail-Notifications
-        </li>
-        <li><strong>Technically mature, data privacy-friendly</strong><br>
-            Open source software, verifiable functionality. No black box.<br>
-            Our servers are all located within the European Union.<br>
-            No external tracking services, advertisements or similar.<br>
-            Standard encrypted transmission.<br>
-            Extensively tested software with hundreds of automated tests as well as operational use during large events
-            with several hundred submitted motions.
-        </li>
-    </ul>
+<section aria-labelledby="motions">
+    <h2 id="motions" class="green">Motions, amendments, resolutions, candidacies</h2>
 
-    <p style="text-align: center; font-weight: bold;">
-        <a href="<?= Html::encode(UrlHelper::createUrl('manager/help')) ?>">
-            <span class="glyphicon glyphicon-chevron-right" aria-hidden="true"></span>
-            Manual / Detailed description of the functionality
-        </a>
-    </p>
-</div>
+    <div class="content infoSite">
+        <p>Antragsgrün allows you to implement your whole motion process, including amendments and candidacies.</p>
+
+        <p><strong>Motions, Resolutions, Statutes, or election programmes</strong> can be submitted and published,
+            either by members or dedicated user groups like the board or eligible delegates.
+            Also submitting <strong>candidacies</strong>, including images and PDF attachments is supported.</p>
+
+        <figure class="homeFigure homeFigureAmendment">
+            <img src="/img/Screenshot-Amendment-en.png" alt="Screenshot of an amendment">
+            <figcaption>
+                Amendments can be shown either separately or (as shown here) in the context of the motion.
+            </figcaption>
+        </figure>
+
+        <p>Besides allowing other members to <strong>comment</strong> on published documents, Antragsgrün encourages
+            constructive proposals by providing an easy-to-use way to <strong>submit amendments</strong> to documents,
+            to be discussed and decided by the board or the member's General Assembly.</p>
+
+        <figure class="homeFigure homeFigurePrint">
+            <img src="/img/Screenshot-Print.png" alt="Druckvorlage">
+            <figcaption>
+                Diverse export options, like a print template
+            </figcaption>
+        </figure>
+
+        <p><strong>Antragsgrün supports organizing assemblies</strong> by many different features:</p>
+
+        <ul>
+            <li>Printing templates can be created automatically<br>(Export as PDF, Spreadsheet or text documents)</li>
+            <li>Accepted amendments can be merged into the original document - creating a final resolution works even with many amendments and ad-hoc changes</li>
+            <li>E-Mail-Notification on relevant events both for administrators and participants</li>
+            <li>Defining responsibilities for motions and topics, internal admin tools for decision-finding</li>
+            <li>... and much more.</li>
+        </ul>
+
+        <p>
+            <a href="https://motion.tools/help" style="font-weight: bold;">
+                <span class="glyphicon glyphicon-chevron-right" aria-hidden="true"></span>
+                Get a more exhaustive list of features
+            </a><br><br>
+            <a href="https://motion.tools/help/member-motion">
+                <span class="glyphicon glyphicon-chevron-right" aria-hidden="true"></span>
+                Tutorial: Submission of motions and resolutions
+            </a><br>
+            <a href="https://motion.tools/help/amendments">
+                <span class="glyphicon glyphicon-chevron-right" aria-hidden="true"></span>
+                Tutorial: Submission of amendments
+            </a><br>
+        </p>
+
+        <br style="clear: both;">
+    </div>
+</section>
+
+<section aria-labelledby="onsite">
+    <h2 id="onsite" class="green">On Site support for conventions</h2>
+
+    <div class="content infoSite">
+        <p>Antragsgrün helps with working with motions and amendments on conventions themselves:</p>
+
+        <ul>
+            <li>An easy administration of the <strong>Agenda</strong></li>
+            <li><strong>Speaking lists</strong></li>
+            <li><strong>Roll Calls and votings</strong> are supported</li>
+            <li>Full-Screen-mode for <strong>projectors</strong> for all relevant content</li>
+        </ul>
+    </div>
+
+
+    <div class="homeFigureTwoHolder">
+        <figure class="homeFigure homeFigureSpeech">
+            <img src="/img/Screenshot-Speaking.png" alt="Screenshot of a speaking list">
+            <figcaption>
+                Flexible administration of a speaking list
+            </figcaption>
+        </figure>
+
+        <figure class="homeFigure homeFigureVoting">
+            <img src="/img/Screenshot-Voting.png" alt="Screenshot of a voting">
+            <figcaption>
+                Votings over motions, amendments and arbitrary decisions
+            </figcaption>
+        </figure>
+    </div>
+</section>
+
+<section aria-labelledby="flexible">
+    <h2 id="flexible" class="green">Flexible and Customizable</h2>
+
+    <div class="content infoSite">
+        <p>Antragsgrün is used by a wide range of organizations is different scenarios and can be customized in many ways
+            to many common settings. For example:</p>
+
+        <ul>
+            <li>The <strong>layout</strong> of the site can be changed on the site itself: the colors, fonts, the logo, and all texts and labels can be customized without effort.</li>
+            <li>The <strong>submission policy</strong> of motions and amendments allows many different variants: setting deadlines, restricting it to eligible user groups, requiring screening by an administrator group, ...</li>
+            <li>Integrating the system into an existing <strong>Single-Sign-On-Mechanism</strong> (like SAML) is possible and can be custom-tailored on request.</li>
+        </ul>
+
+        <p>For further features and organization-specific adaptions we <strong>professional support.</strong></p>
+
+        <p><strong>Contact for support</strong>:<br>
+            Tobias Hößl<br>
+            <a href="mailto:info@antragsgruen.de">info@antragsgruen.de</a>
+    </div>
+</section>
+
+<section aria-labelledby="mature">
+    <h2 id="mature" class="green">Mature, Open, Privacy friendly</h2>
+
+    <div class="content infoSite">
+        <p>Antragsgrün is actively used on political conventions and assemblies <strong>since more than ten years</strong>
+            and helps both small informal working group as well as large general assemblies with hundreds of delegates
+            and thousands of amendments.</p>
+
+        <p>Antragsgrün is continuously developed as <strong>Open Source</strong> (AGPL) in cooperation with
+            organizations using it. Antragsgrün can, on the one side, be downloaded, installed and used for free as it is.
+            The download and an installation guide can be found on <a href="https://github.com/CatoTH/antragsgruen">Github</a>.
+
+        <p>On the other side, we also provide <strong>professional support</strong>,
+            hosting and the implementation of new features and organization-specific changes on a per-request basis.</p>
+
+        <p>We care a lot about data privacy: we do not collect unnecessary personal data, we <strong>do not use trackers</strong>
+            and don't do Ads. Our main servers are located in the European Union, hosting in the U.S. is available on request.</p>
+    </div>
+</section>
 
 <h2 id="create_version" class="green">Use it</h2>
 <div class="content">
     <p>Discuss.green / Antragsgrün is open source software and can be installed on any web server that supports PHP and MySQL.</p>
     <p>
-        For european green parties, we provide <strong>free hosting</strong> of sites on this domain.
+        For European and Canadian Green parties, we provide <strong>free hosting</strong> of sites on this domain.
         You can start using or evaluating this tool by answering a couple of questions about your use case and providing
-        a valid e-mail-address. Creating your own instance only takes two minutes.
+        a valid e-mail-address. Creating your own instance only takes about two minutes.
     </p>
 
     <div class="downloadCreate">
@@ -137,9 +246,9 @@ $params = \app\models\settings\AntragsgruenApp::getInstance();
     </div>
 
     <p>
-        <strong>About internationalization:</strong> We provide an english and german version,
-        as well as a nearly-completele french version. If you are interested in helping translate this tool
-        in other languages, please contact us.
+        <strong>About internationalization:</strong> We provide an English and German version,
+        as well as nearly-completele versions in French and other community-contributed languages.
+        If you are interested in helping translate this tool in other languages, please contact us.
     </p>
     <p>
         <a href="<?= Html::encode(UrlHelper::createUrl('manager/free-hosting')) ?>">
@@ -161,6 +270,5 @@ $params = \app\models\settings\AntragsgruenApp::getInstance();
     </p>
     <ul>
         <li>E-Mail: <a href="mailto:info@discuss.green">info@discuss.green</a></li>
-        <li>Phone: +49-1515-6024223</li>
     </ul>
 </div>